# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /backtest/tasks:
    post:
      description: Creates a new backtest task with the specified parameters. The task will be queued for processing.
      operationId: BacktestController_createTask
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBacktestTaskDto'
      responses:
        '201':
          description: Backtest task created successfully
        '400':
          description: Invalid input or date range
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: &ref_1
        - JWT-auth: []
      summary: Create a new backtest task
      tags: &ref_2
        - Backtest
    get:
      description: Returns a paginated list of backtest tasks for the authenticated user. By default, returns only non-archived and non-deleted tasks.
      operationId: BacktestController_listTasks
      parameters:
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination
          schema:
            example: cm1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: 'Number of tasks per page (default: 20, max: 100)'
          schema:
            minimum: 1
            maximum: 100
            example: 20
            type: number
        - name: status
          required: false
          in: query
          description: Filter by task status
          schema:
            example: COMPLETED
            type: string
            enum:
              - PENDING
              - QUEUED
              - RUNNING
              - COMPLETED
              - FAILED
              - CANCELLED
        - name: archived
          required: false
          in: query
          description: 'Filter by archive status (default: false - shows only non-archived tasks)'
          schema:
            example: false
            type: boolean
      responses:
        '200':
          description: Tasks retrieved successfully
        '401':
          description: Unauthorized
      security: *ref_1
      summary: List all backtest tasks
      tags: *ref_2
  /backtest/tasks/{id}:
    get:
      description: Returns a specific backtest task including its result summary (without individual trades). Archived tasks are still accessible. Returns 404 only if task is soft deleted or not found.
      operationId: BacktestController_getTaskById
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest task ID
          schema:
            type: string
      responses:
        '200':
          description: Task retrieved successfully (including archived tasks)
        '401':
          description: Unauthorized
        '404':
          description: Task not found or has been deleted
      security: *ref_1
      summary: Get backtest task by ID
      tags: *ref_2
    delete:
      description: Soft deletes a backtest task. The task will be permanently hidden from all endpoints. Cannot delete a task that is currently running.
      operationId: BacktestController_deleteTask
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest task ID
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '400':
          description: Cannot delete a running task
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security: *ref_1
      summary: Soft delete a backtest task
      tags: *ref_2
  /backtest/tasks/{id}/archive:
    patch:
      description: Archives a backtest task. Cannot archive a task that is currently running.
      operationId: BacktestController_archiveTask
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest task ID
          schema:
            type: string
      responses:
        '200':
          description: Task archived successfully
        '400':
          description: Cannot archive a running task
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security: *ref_1
      summary: Archive a backtest task
      tags: *ref_2
  /backtest/tasks/{id}/unarchive:
    patch:
      description: Restores an archived backtest task to active status.
      operationId: BacktestController_unarchiveTask
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest task ID
          schema:
            type: string
      responses:
        '200':
          description: Task unarchived successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security: *ref_1
      summary: Unarchive a backtest task
      tags: *ref_2
  /backtest/results/{resultId}/trades:
    get:
      description: Returns individual trades from a backtest result with cursor-based pagination and filtering options. Use this endpoint to retrieve large datasets efficiently.
      operationId: BacktestController_getResultTrades
      parameters:
        - name: resultId
          required: true
          in: path
          description: Backtest result ID
          schema:
            type: string
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination (last trade ID from previous page)
          schema:
            example: cm1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: Number of trades to return per page
          schema:
            minimum: 1
            maximum: 1000
            default: 100
            example: 100
            type: number
        - name: symbol
          required: false
          in: query
          description: Filter by symbol
          schema:
            example: BTCUSDT
            type: string
        - name: timeframe
          required: false
          in: query
          description: Filter by timeframe
          schema:
            example: 4h
            type: string
        - name: profitable
          required: false
          in: query
          description: Filter by profitability (true = winning trades, false = losing trades)
          schema:
            example: true
            type: boolean
        - name: exit_reason
          required: false
          in: query
          description: Filter by exit reason
          schema:
            example: stop_loss_hit
            type: string
            enum:
              - stop_loss_hit
              - take_profit_hit
              - pattern_expired
      responses:
        '200':
          description: Trades retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestTradesResponseDto'
        '401':
          description: Unauthorized
        '404':
          description: Result not found
      security: *ref_1
      summary: Get backtest result trades with pagination
      tags: *ref_2
  /backtest/trades/{tradeId}:
    get:
      description: Returns detailed information about a specific backtest trade including 50 candles before and 50 candles after the first top/bottom from the pattern. This endpoint fetches candle data from the GraphQL chart data service.
      operationId: BacktestController_getTradeDetails
      parameters:
        - name: tradeId
          required: true
          in: path
          description: Backtest trade ID
          schema:
            type: string
      responses:
        '200':
          description: Trade details retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Trade not found
      security: *ref_1
      summary: Get backtest trade details with candle data
      tags: *ref_2
  /backtest/strategies:
    post:
      description: Creates a new backtest strategy template that can be reused for multiple backtest tasks.
      operationId: BacktestController_createStrategy
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBacktestStrategyDto'
      responses:
        '201':
          description: Strategy created successfully
        '400':
          description: Invalid input or duplicate strategy name
        '401':
          description: Unauthorized
      security: *ref_1
      summary: Create a new backtest strategy
      tags: *ref_2
    get:
      description: Returns a paginated list of backtest strategies for the authenticated user. By default, returns only non-archived and non-deleted strategies.
      operationId: BacktestController_listStrategies
      parameters:
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination
          schema:
            example: cm1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: 'Number of strategies per page (default: 20, max: 100)'
          schema:
            minimum: 1
            maximum: 100
            example: 20
            type: number
        - name: archived
          required: false
          in: query
          description: 'Filter by archive status (default: false - shows only non-archived strategies)'
          schema:
            example: false
            type: boolean
      responses:
        '200':
          description: Strategies retrieved successfully
        '401':
          description: Unauthorized
      security: *ref_1
      summary: List all backtest strategies
      tags: *ref_2
  /backtest/strategies/{id}:
    get:
      description: Returns a specific backtest strategy. Archived strategies are still accessible. Returns 404 only if strategy is soft deleted or not found.
      operationId: BacktestController_getStrategyById
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy retrieved successfully (including archived strategies)
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found or has been deleted
      security: *ref_1
      summary: Get backtest strategy by ID
      tags: *ref_2
    delete:
      description: Soft deletes a backtest strategy. The strategy will be permanently hidden from all endpoints.
      operationId: BacktestController_deleteStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: *ref_1
      summary: Soft delete a backtest strategy
      tags: *ref_2
  /backtest/strategies/{id}/archive:
    patch:
      description: Archives a backtest strategy.
      operationId: BacktestController_archiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy archived successfully
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: *ref_1
      summary: Archive a backtest strategy
      tags: *ref_2
  /backtest/strategies/{id}/unarchive:
    patch:
      description: Restores an archived backtest strategy to active status.
      operationId: BacktestController_unarchiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Backtest strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy unarchived successfully
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: *ref_1
      summary: Unarchive a backtest strategy
      tags: *ref_2
...
  /strategies:
    get:
      description: Returns paginated list of strategies for the authenticated user with optional filters
      operationId: StrategyController_listStrategies
      parameters:
        - name: trading_account_id
          required: false
          in: query
          description: Filter by trading account ID
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: mode
          required: false
          in: query
          description: Filter by trading mode
          schema:
            type: string
            enum:
              - REAL
              - PAPER
        - name: is_live
          required: false
          in: query
          description: Filter by live status
          schema:
            example: true
            type: boolean
        - name: include_archived
          required: false
          in: query
          description: Include archived strategies in results
          schema:
            default: false
            example: false
            type: boolean
        - name: archived_only
          required: false
          in: query
          description: Show only archived strategies
          schema:
            default: false
            example: false
            type: boolean
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination (strategy ID)
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Strategies retrieved successfully
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: List user strategies
      tags: &ref_5
        - Strategies
    post:
      description: Creates a new trading strategy and an initial portfolio. The portfolio type matches the strategy mode (PAPER by default).
      operationId: StrategyController_createStrategy
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyDto'
      responses:
        '201':
          description: Strategy created successfully
        '401':
          description: Unauthorized
        '404':
          description: Trading account not found
        '409':
          description: Strategy with this name already exists for this account
      security:
        - JWT-auth: []
      summary: Create new strategy
      tags: *ref_5
...
  /strategies/{id}:
    get:
      description: Returns detailed information about a strategy including portfolios and trading account. User must own the strategy OR strategy must be public.
      operationId: StrategyController_getStrategyById
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy retrieved successfully
        '401':
          description: Unauthorized
        '403':
          description: You do not have access to this strategy
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Get strategy details
      tags: *ref_5
    patch:
      description: Updates strategy configuration. Cannot change mode while is_live=true. Cannot switch to REAL mode without a REAL portfolio.
      operationId: StrategyController_updateStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyDto'
      responses:
        '200':
          description: Strategy updated successfully
        '400':
          description: Invalid update - check validation rules
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
        '409':
          description: Strategy with this name already exists for this account
      security:
        - JWT-auth: []
      summary: Update strategy settings
      tags: *ref_5
    delete:
      description: Deletes a strategy. Cannot delete if is_live=true or if there are open positions. This action is permanent.
      operationId: StrategyController_deleteStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy deleted successfully
        '400':
          description: Cannot delete - strategy is live or has open positions
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Delete strategy
      tags: *ref_5
...
  /strategies/{id}/archive:
    post:
      description: Archives a strategy. Cannot archive if is_live=true. Archived strategies are hidden by default but can be filtered in list endpoints.
      operationId: StrategyController_archiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy archived successfully
        '400':
          description: Cannot archive - strategy is live or already archived
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Archive strategy
      tags: *ref_5
  /strategies/{id}/unarchive:
    post:
      description: Unarchives a strategy. Strategy must be archived.
      operationId: StrategyController_unarchiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy unarchived successfully
        '400':
          description: Cannot unarchive - strategy is not archived
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Unarchive strategy
      tags: *ref_5
...
info:
  title: Crypto Watcher Bot API
  description: API for crypto trading signal bot with automated order execution
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    CreateBacktestTaskDto:
      type: object
      properties:
        name:
          type: string
          description: User-friendly name for the backtest
          example: Double Top Strategy - Q1 2024
        description:
          type: string
          description: Detailed description of the backtest
          example: Testing double top pattern across major pairs
        strategy_id:
          type: string
          description: Optional strategy template ID to use
          example: cm1234567890abcdefghijk
        priority:
          type: string
          description: Queue priority for the backtest task
          enum:
            - high
            - normal
            - low
          default: normal
          example: normal
        symbols:
          description: List of symbols to backtest
          example:
            - BTCUSDT
            - ETHUSDT
            - SOLUSDT
          type: array
          items:
            type: string
        timeframes:
          description: List of timeframes to backtest
          example:
            - 4h
            - 1h
          type: array
          items:
            type: string
        start_date:
          type: string
          description: Start date for backtest (ISO 8601 format)
          example: '2024-01-01T00:00:00.000Z'
        end_date:
          type: string
          description: End date for backtest (ISO 8601 format)
          example: '2024-12-31T23:59:59.999Z'
        is_combine_timeframe:
          type: boolean
          description: If true, create separate backtest tasks for each timeframe. If false, combine all timeframes in a single task.
          example: false
          default: false
        is_combine_symbol:
          type: boolean
          description: If true, create separate backtest tasks for each symbol. If false, combine all symbols in a single task.
          example: false
          default: false
        strategy_config:
          description: Strategy configuration for the backtest
          examples:
            percentage_mode:
              summary: Percentage Risk Mode
              value:
                strategy_type: pattern_breakout
                allowed_patterns:
                  - double_top
                  - double_bottom
                risk_mode: PERCENTAGE
                risk_per_trade_percentage: 5
                initial_capital: 10000
                min_risk_ratio: 1.5
                position_type: long
            fixed_amount_mode:
              summary: Fixed Amount Risk Mode
              value:
                strategy_type: pattern_breakout
                allowed_patterns:
                  - double_top
                  - double_bottom
                  - head_shoulders
                risk_mode: FIXED_AMOUNT
                fixed_risk_amount: 100
                initial_capital: 10000
                max_risk_ratio: 5
          allOf:
            - $ref: '#/components/schemas/BacktestStrategyConfigDto'
      required:
        - name
        - symbols
        - timeframes
        - start_date
        - end_date
        - strategy_config
...
    BacktestTradesResponseDto:
      type: object
      properties:
        data:
          type: array
          description: Array of backtest trades
        pagination:
          description: Pagination metadata
          allOf:
            - $ref: '#/components/schemas/PaginationMetaDto'
        summary:
          description: Summary statistics for filtered trades
          allOf:
            - $ref: '#/components/schemas/TradesSummaryDto'
      required:
        - data
        - pagination
        - summary
    CreateBacktestStrategyDto:
      type: object
      properties:
        name:
          type: string
          description: User-friendly name for the strategy
          example: Aggressive Double Top Strategy
        strategy_type:
          type: string
          description: Strategy type identifier
          example: pattern_breakout
        allowed_patterns:
          description: List of allowed pattern types to trade
          example:
            - double_top
            - double_bottom
          type: array
          items:
            type: string
        allowed_symbols:
          description: List of allowed symbols (empty array or omit for all symbols)
          example:
            - BTCUSDT
            - ETHUSDT
          type: array
          items:
            type: string
        timeframes:
          description: List of allowed timeframes (empty array or omit for all timeframes)
          example:
            - 4h
            - 1h
          type: array
          items:
            type: string
        position_type:
          type: string
          description: Position type filter (LONG, SHORT, or omit for both)
          enum:
            - LONG
            - SHORT
          example: LONG
        config:
          description: Strategy configuration including risk management
          allOf:
            - $ref: '#/components/schemas/BacktestStrategyConfigDto'
      required:
        - name
        - strategy_type
        - allowed_patterns
        - config
...